Resolved CID 351543 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument state->getDelta() to format specifier %u was expected to have type unsigned int but has type long long

+++ src/world/Chat/Commands/DebugCommands.cpp
@@ -149,7 +149,7 @@ bool ChatHandler::HandleSetOrientationCommand(const char* args, WorldSession* se
 bool ChatHandler::HandleDebugDumpState(const char* /*args*/, WorldSession* session)
 {
     auto state = ServerState::instance();
-    SystemMessage(session, "Delta: %u", state->getDelta());
+    SystemMessage(session, "Delta: %u", static_cast<uint32_t>(state->getDelta()));
     return true;
 }
